Refactor: give the payment instructions and CopyButton a home of their own
Preparatory only, no behaviour change. Two things the membership page is about
to need were sitting inside single callers:

- `getInvoiceInstructions` read `club_settings` from inside
  `membership-email.server.ts`, whose header said the club's bank details were
  never returned to the browser. That stops being true the moment a member can
  read them off /membership, so the read moves to `club-settings.server.ts` and
  its tests move with it.
- The copy-to-clipboard button was defined inside `/manager/api-tokens`. This is
  the third place in the app that copies a string, so it becomes
  `components/site/CopyButton.tsx` with the failure path documented.
